refactor: move constants to common and r/wers to io
