+++ src/decoding/frame.rs
@@ -1,12 +1,92 @@
+use crate::common::{MAGIC_NUM, MAX_WINDOW_SIZE, MIN_WINDOW_SIZE};
 use crate::decoding::errors::{FrameDescriptorError, FrameHeaderError, ReadFrameHeaderError};
-use crate::io::Read;
+use crate::io_std::Read;
 
-/// This magic number is included at the start of a single Zstandard frame
-pub const MAGIC_NUM: u32 = 0xFD2F_B528;
-/// The minimum window size is defined as 1 KB
-pub const MIN_WINDOW_SIZE: u64 = 1024;
-/// The maximum window size is 3.75TB
-pub const MAX_WINDOW_SIZE: u64 = (1 << 41) + 7 * (1 << 38);
+/// Read a single serialized frame from the reader and return a tuple containing the parsed frame and the number of bytes read.
+pub fn read_frame_header(mut r: impl Read) -> Result<(Frame, u8), ReadFrameHeaderError> {
+    use ReadFrameHeaderError as err;
+    let mut buf = [0u8; 4];
+
+    r.read_exact(&mut buf).map_err(err::MagicNumberReadError)?;
+    let mut bytes_read = 4;
+    let magic_num = u32::from_le_bytes(buf);
+
+    // Skippable frames have a magic number in this interval
+    if (0x184D2A50..=0x184D2A5F).contains(&magic_num) {
+        r.read_exact(&mut buf)
+            .map_err(err::FrameDescriptorReadError)?;
+        let skip_size = u32::from_le_bytes(buf);
+        return Err(ReadFrameHeaderError::SkipFrame {
+            magic_number: magic_num,
+            length: skip_size,
+        });
+    }
+
+    if magic_num != MAGIC_NUM {
+        return Err(ReadFrameHeaderError::BadMagicNumber(magic_num));
+    }
+
+    r.read_exact(&mut buf[0..1])
+        .map_err(err::FrameDescriptorReadError)?;
+    let desc = FrameDescriptor(buf[0]);
+
+    bytes_read += 1;
+
+    let mut frame_header = FrameHeader {
+        descriptor: FrameDescriptor(desc.0),
+        dict_id: None,
+        frame_content_size: 0,
+        window_descriptor: 0,
+    };
+
+    if !desc.single_segment_flag() {
+        r.read_exact(&mut buf[0..1])
+            .map_err(err::WindowDescriptorReadError)?;
+        frame_header.window_descriptor = buf[0];
+        bytes_read += 1;
+    }
+
+    let dict_id_len = desc.dictionary_id_bytes()? as usize;
+    if dict_id_len != 0 {
+        let buf = &mut buf[..dict_id_len];
+        r.read_exact(buf).map_err(err::DictionaryIdReadError)?;
+        bytes_read += dict_id_len;
+        let mut dict_id = 0u32;
+
+        #[allow(clippy::needless_range_loop)]
+        for i in 0..dict_id_len {
+            dict_id += (buf[i] as u32) << (8 * i);
+        }
+        if dict_id != 0 {
+            frame_header.dict_id = Some(dict_id);
+        }
+    }
+
+    let fcs_len = desc.frame_content_size_bytes()? as usize;
+    if fcs_len != 0 {
+        let mut fcs_buf = [0u8; 8];
+        let fcs_buf = &mut fcs_buf[..fcs_len];
+        r.read_exact(fcs_buf)
+            .map_err(err::FrameContentSizeReadError)?;
+        bytes_read += fcs_len;
+        let mut fcs = 0u64;
+
+        #[allow(clippy::needless_range_loop)]
+        for i in 0..fcs_len {
+            fcs += (fcs_buf[i] as u64) << (8 * i);
+        }
+        if fcs_len == 2 {
+            fcs += 256;
+        }
+        frame_header.frame_content_size = fcs;
+    }
+
+    let frame: Frame = Frame {
+        header: frame_header,
+    };
+
+    Ok((frame, bytes_read as u8))
+}
 
 /// Zstandard compressed data is made of one or more [Frame]s. Each frame is independent and can be
 /// decompressed independently of other frames.
@@ -175,90 +255,4 @@ impl FrameDescriptor {
             other => Err(FrameDescriptorError::InvalidFrameContentSizeFlag { got: other }),
         }
     }
-}
-
-/// Read a single serialized frame from the reader and return a tuple containing the parsed frame and the number of bytes read.
-pub fn read_frame_header(mut r: impl Read) -> Result<(Frame, u8), ReadFrameHeaderError> {
-    use ReadFrameHeaderError as err;
-    let mut buf = [0u8; 4];
-
-    r.read_exact(&mut buf).map_err(err::MagicNumberReadError)?;
-    let mut bytes_read = 4;
-    let magic_num = u32::from_le_bytes(buf);
-
-    // Skippable frames have a magic number in this interval
-    if (0x184D2A50..=0x184D2A5F).contains(&magic_num) {
-        r.read_exact(&mut buf)
-            .map_err(err::FrameDescriptorReadError)?;
-        let skip_size = u32::from_le_bytes(buf);
-        return Err(ReadFrameHeaderError::SkipFrame {
-            magic_number: magic_num,
-            length: skip_size,
-        });
-    }
-
-    if magic_num != MAGIC_NUM {
-        return Err(ReadFrameHeaderError::BadMagicNumber(magic_num));
-    }
-
-    r.read_exact(&mut buf[0..1])
-        .map_err(err::FrameDescriptorReadError)?;
-    let desc = FrameDescriptor(buf[0]);
-
-    bytes_read += 1;
-
-    let mut frame_header = FrameHeader {
-        descriptor: FrameDescriptor(desc.0),
-        dict_id: None,
-        frame_content_size: 0,
-        window_descriptor: 0,
-    };
-
-    if !desc.single_segment_flag() {
-        r.read_exact(&mut buf[0..1])
-            .map_err(err::WindowDescriptorReadError)?;
-        frame_header.window_descriptor = buf[0];
-        bytes_read += 1;
-    }
-
-    let dict_id_len = desc.dictionary_id_bytes()? as usize;
-    if dict_id_len != 0 {
-        let buf = &mut buf[..dict_id_len];
-        r.read_exact(buf).map_err(err::DictionaryIdReadError)?;
-        bytes_read += dict_id_len;
-        let mut dict_id = 0u32;
-
-        #[allow(clippy::needless_range_loop)]
-        for i in 0..dict_id_len {
-            dict_id += (buf[i] as u32) << (8 * i);
-        }
-        if dict_id != 0 {
-            frame_header.dict_id = Some(dict_id);
-        }
-    }
-
-    let fcs_len = desc.frame_content_size_bytes()? as usize;
-    if fcs_len != 0 {
-        let mut fcs_buf = [0u8; 8];
-        let fcs_buf = &mut fcs_buf[..fcs_len];
-        r.read_exact(fcs_buf)
-            .map_err(err::FrameContentSizeReadError)?;
-        bytes_read += fcs_len;
-        let mut fcs = 0u64;
-
-        #[allow(clippy::needless_range_loop)]
-        for i in 0..fcs_len {
-            fcs += (fcs_buf[i] as u64) << (8 * i);
-        }
-        if fcs_len == 2 {
-            fcs += 256;
-        }
-        frame_header.frame_content_size = fcs;
-    }
-
-    let frame: Frame = Frame {
-        header: frame_header,
-    };
-
-    Ok((frame, bytes_read as u8))
 }
