+++ src/decoding/mod.rs
@@ -7,10 +7,8 @@ mod streaming_decoder;
 pub use frame_decoder::{BlockDecodingStrategy, FrameDecoder};
 pub use streaming_decoder::StreamingDecoder;
 
-pub(crate) mod bit_reader;
-pub(crate) mod bit_reader_reverse;
 pub(crate) mod block_decoder;
-pub(crate) mod decodebuffer;
+pub(crate) mod decode_buffer;
 pub(crate) mod dictionary;
 pub(crate) mod frame;
 pub(crate) mod literals_section_decoder;
