+++ src/decoding/sequence_section_decoder.rs
@@ -1,13 +1,13 @@
 use super::super::blocks::sequence_section::ModeType;
 use super::super::blocks::sequence_section::Sequence;
 use super::super::blocks::sequence_section::SequencesHeader;
-use super::bit_reader_reverse::BitReaderReversed;
 use super::scratch::FSEScratch;
 use crate::blocks::sequence_section::{
     MAX_LITERAL_LENGTH_CODE, MAX_MATCH_LENGTH_CODE, MAX_OFFSET_CODE,
 };
 use crate::decoding::errors::DecodeSequenceError;
 use crate::fse::FSEDecoder;
+use crate::io::BitReaderReversed;
 use alloc::vec::Vec;
 
 /// Decode the provided source as a series of sequences into the supplied `target`.
