+++ src/encoding/blocks/compressed.rs
@@ -2,9 +2,10 @@ use alloc::vec::Vec;
 
 use crate::{
     encoding::frame_compressor::CompressState,
-    encoding::{bit_writer::BitWriter, Matcher, Sequence},
+    encoding::{Matcher, Sequence},
     fse::fse_encoder::{default_ll_table, default_ml_table, default_of_table, FSETable, State},
     huff0::huff0_encoder,
+    io::BitWriter,
 };
 
 pub fn compress_block<M: Matcher>(state: &mut CompressState<M>, output: &mut Vec<u8>) {
