+++ src/fse/mod.rs
@@ -78,7 +78,7 @@ fn roundtrip() {
 /// Asserts that the decoded data equals the input
 #[cfg(any(test, feature = "fuzz_exports"))]
 pub fn round_trip(data: &[u8]) {
-    use crate::{decoding::bit_reader_reverse::BitReaderReversed, encoding::bit_writer::BitWriter};
+    use crate::io::{BitReaderReversed, BitWriter};
     use fse_encoder::FSEEncoder;
 
     if data.len() < 2 {
