+++ src/huff0/mod.rs
@@ -13,7 +13,7 @@ pub mod huff0_encoder;
 /// Asserts that the decoded data equals the input
 #[cfg(any(test, feature = "fuzz_exports"))]
 pub fn round_trip(data: &[u8]) {
-    use crate::{decoding::bit_reader_reverse::BitReaderReversed, encoding::bit_writer::BitWriter};
+    use crate::io::{BitReaderReversed, BitWriter};
     use alloc::vec::Vec;
 
     if data.len() < 2 {
