+++ src/io/mod.rs
@@ -1,0 +1,9 @@
+//! Encoding agnostic ways to read and write binary data
+
+mod bit_reader;
+mod bit_reader_reverse;
+mod bit_writer;
+
+pub(crate) use bit_reader::*;
+pub(crate) use bit_reader_reverse::*;
+pub(crate) use bit_writer::*;
