+++ src/lib.rs
@@ -32,8 +32,10 @@ macro_rules! vprintln {
     }
 }
 
+mod common;
 pub mod decoding;
 pub mod encoding;
+mod io;
 
 pub(crate) mod blocks;
 
@@ -50,7 +52,7 @@ pub(crate) mod huff0;
 mod tests;
 
 #[cfg(feature = "std")]
-pub mod io;
+pub mod io_std;
 
 #[cfg(not(feature = "std"))]
 pub mod io_nostd;
